+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### 🚜 Refactor
+
+- Removing caddy group add
+
 ## [0.21.3] - 2026-02-12
 
 ### 🐛 Bug Fixes